+++ src/stats-aggregator.ts
@@ -1,5 +1,4 @@
 import StatsGatherer, { GetStatsEvent, StatsEvent } from "webrtc-stats-gatherer";
-import { v4 as uuidv4 } from "uuid";
 
 import GenesysCloudWebrtSdk, { IConversationParticipantFromEvent, IExtendedMediaSession } from ".";
 
@@ -10,7 +9,6 @@ interface ISentStats {
 }
 
 export class StatsAggregator {
-  private mediaResourceId: string;
   private statsGatherer?: StatsGatherer;
 
   private setBaseline = false;
@@ -24,8 +22,6 @@ export class StatsAggregator {
   private boundStatsHandler?: (stats: StatsEvent) => void;
 
   constructor(private session: IExtendedMediaSession, private sdk: GenesysCloudWebrtSdk) {
-    this.mediaResourceId = uuidv4();
-
     if (this.shouldGatherImmediately(session)) {
       this.startGatheringStats();
     }
@@ -225,13 +221,12 @@ export class StatsAggregator {
       to: this.session.peerID,
       genesysWebrtc: {
         jsonrpc: '2.0',
-        id: uuidv4(),
+        id: globalThis.crypto.randomUUID(),
         method: 'report-statistics',
         params: {
           sessionId: this.session.id,
           conversationId,
           communicationId,
-          mediaResourceId: this.mediaResourceId,
           stats: statsData
         }
       }